:art: refactoring the code
