+++ pyecharts/constants.py
@@ -2,11 +2,29 @@
 # coding=utf-8
 from __future__ import unicode_literals
 
+import os
+import json
+import codecs
+from pyecharts.utils import get_resource_dir
+
+
 DEFAULT_HOST = 'https://chfw.github.io/jupyter-echarts/echarts'
 
 CONFIGURATION = dict(
     HOST='/nbextensions/echarts'
 )
+
+DEFAULT_ECHARTS_REGISTRY = os.path.join(
+    get_resource_dir('templates'), 'js', 'echarts', 'registry.json')
+
+with codecs.open(DEFAULT_ECHARTS_REGISTRY, 'r', 'utf-8') as f:
+    content = f.read()
+    CONFIG = json.loads(content)
+
+DEFAULT_JS_LIBRARIES = CONFIG['FILE_MAP']
+
+CITY_NAME_PINYIN_MAP = CONFIG['PINYIN_MAP']
+
 
 CITY_GEO_COORDS = {
     '阿城': [126.58, 45.32],
