+++ pyecharts/template.py
@@ -8,28 +8,8 @@ import sys
 import codecs
 from jinja2 import Environment, FileSystemLoader
 import pyecharts.constants as constants
-import json
+from pyecharts.utils import get_resource_dir
 
-
-def get_resource_dir(folder):
-    """
-
-    :param folder:
-    :return:
-    """
-    current_path = os.path.dirname(__file__)
-    resource_path = os.path.join(current_path, folder)
-    return resource_path
-
-
-#with codecs.open(os.path.join(get_resource_dir('templates'), 'js', 'echarts', 'registry.json'), 'r', 'utf-8') as f:
-with open(os.path.join(get_resource_dir('templates'), 'js', 'echarts', 'registry.json'), 'rb') as f:
-    content = f.read().decode('utf-8')
-    CONFIG = json.loads(content)
-
-DEFAULT_JS_LIBRARIES = CONFIG['FILE_MAP']
-
-CITY_NAME_PINYIN_MAP = CONFIG['PINYIN_MAP']
 
 PY2 = sys.version_info[0] == 2
 
@@ -107,7 +87,7 @@ def produce_require_configuration(dependencies, jshost):
     require_conf_items = [
         "'%s': '%s/%s'" % (key,
                            jshost,
-                           DEFAULT_JS_LIBRARIES.get(key, key))
+                           constants.DEFAULT_JS_LIBRARIES.get(key, key))
         for key in _d]
     require_libraries = ["'%s'" % key for key in _d]
     return dict(
@@ -124,7 +104,7 @@ def produce_html_script_list(dependencies):
     """
     _d = ensure_echarts_is_in_the_front(dependencies)
     script_list = [
-        '%s' % DEFAULT_JS_LIBRARIES.get(key, key)
+        '%s' % constants.DEFAULT_JS_LIBRARIES.get(key, key)
         for key in _d]
     return script_list
