+++ pyecharts/utils.py
@@ -1,0 +1,12 @@
+import os
+
+
+def get_resource_dir(folder):
+    """
+
+    :param folder:
+    :return:
+    """
+    current_path = os.path.dirname(__file__)
+    resource_path = os.path.join(current_path, folder)
+    return resource_path
