+++ test.sh
@@ -1,3 +1,3 @@
 cd test
 nosetests --with-coverage --cover-package pyecharts --cover-package test
-mv .coverage ../
+
